Add docstrings across src/mlframe/reporting/** (mass interrogate-coverage sweep)
Real, substantive module/class/function/method docstrings. No code logic
changed -- docstrings only, plus RUF100/RUF046 mechanical ruff --fix passes.

+++ src/mlframe/reporting/charts/confusion_matrix_plot.py
@@ -187,7 +187,7 @@ def plot_confusion_matrix(
                     color = auto_text_color(val if np.isfinite(val) else vmin, cmap, vmin=vmin, vmax=vmax)
                     # Raw-count cells format as ints ('d'); the display matrix is float64 either way, so cast the
                     # value to int for an integer format spec (sklearn prints counts as ints too).
-                    cell_val = int(round(val)) if values_format.endswith(("d", "n")) else display[i, j]
+                    cell_val = round(val) if values_format.endswith(("d", "n")) else display[i, j]
                     ax.text(j, i, format(cell_val, values_format), ha="center", va="center", fontsize=8, color=color)
 
     if colorbar:
